requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
mysql-connector-python==9.3.0
pillow==11.3.0
platformdirs==4.3.7
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.30.0
virtualenvwrapper-win==1.2.7
